feat(ContextConfigurationComponent): support all input types
