+++ docs/.vuepress/stores/options.js
@@ -8,97 +8,97 @@ export const useOptionsStore = defineStore("color", {
         group: "Color",
         value: { default: "199" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_ASSUME_UNCHANGED: {
         group: "Color",
         value: { default: "81" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_FAILED: {
         group: "Color",
         value: { default: "88" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_HEAD: {
         group: "Color",
         value: { default: "140" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_HOST: {
         group: "Color",
         value: { default: "109" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_INACTIVE: {
         group: "Color",
         value: { default: "247" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_PUSH_REMOTE: {
         group: "Color",
         value: { default: "111" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_REMOTE: {
         group: "Color",
         value: { default: "216" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_SKIP_WORKTREE: {
         group: "Color",
         value: { default: "81" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_STAGED: {
         group: "Color",
         value: { default: "120" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_STASH: {
         group: "Color",
         value: { default: "81" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_SUCCEEDED: {
         group: "Color",
         value: { default: "76" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_TAG: {
         group: "Color",
         value: { default: "86" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_UNSTAGED: {
         group: "Color",
         value: { default: "162" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_USER: {
         group: "Color",
         value: { default: "109" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_COLOR_WORKDIR: {
         group: "Color",
         value: { default: "39" },
         notes: "ANSI or hex",
-        type: "string",
+        type: "color",
       },
       GIT_PROMPT_KIT_HIDE_INACTIVE_AHEAD_BEHIND: {
         group: "Behavior",
